改进unpacker的consume触发机制
避免重复触发consume，以避免可能出现调用堆栈过深。
